Vulcan effects, Vulcan in campaign mechanics scene.

+++ Project -v1.0.2 - 4.2.0/Project -v1.0.2 - 4.2.0/Assets/BloodMist.cs
@@ -19,7 +19,7 @@ public class BloodMist : TargetAbility {
 			}
 		}
 	}
-	public float maxChargeCount;
+	public int maxChargeCount;
 
 
 	override
